+++ FrontEnd/src/components/mypage/MyRecipe.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+const MyRecipe = () => {
+  return;
+  <>MyRecipy</>;
+};
+
+export default MyRecipe;
